Update code for calculating risk parameters & add text folder
NOTE: In previous versions, the risk coefficients are calculated by choosing the minimum distance between two pairs of coefficients (sigma and alpha) for two questions. However, the EXECUTION WAS WRONG as the distance was only done row-wise, meaning only comparing between sigma1 and sigma2 for the same alpha.

The code is now updated to find the distance between one pair of alpha1 and sigma1 with any other pairs of alpha2 and sigma2.

CAVEATS: computational time is gargantuan (several hours) even though the grid space is already sparse